DeleteMessageMixin fixes
Changed post override to delete override.  Overriding post and calling the super delete method was causing the deletion to happen no matter how one would override the Delete method.
Added logic to only actually delete if the request is not an ajax request.  This prevents the parent deletion being called twice and throwing a 404 error on get_object when it is called on the second POST.